Moved  mermaid related code to a dedicated feature dir. `vercelChat/mermaid`
