+++ components/VercelChat/ToolComponents.tsx
@@ -1,9 +1,9 @@
 import { ImageSkeleton } from "@/components/ui/ImageSkeleton";
 import { ImageResult } from "@/components/ui/ImageResult";
 import { ImageGenerationResult } from "@/lib/tools/generateImage";
-import MermaidDiagram from "../Chat/mermaid/MermaidDiagram";
+import MermaidDiagram from "@/components/VercelChat/tools/mermaid/MermaidDiagram";
+import { MermaidDiagramSkeleton } from "@/components/VercelChat/tools/mermaid/MermaidDiagramSkeleton";
 import { GenerateMermaidDiagramResult } from "@/lib/tools/generateMermaidDiagram";
-import { MermaidDiagramSkeleton } from "../ui/MermaidDiagramSkeleton";
 import CreateArtistToolCall from "./tools/CreateArtistToolCall";
 import CreateArtistToolResult from "./tools/CreateArtistToolResult";
 import { CreateArtistResult } from "@/lib/tools/createArtist";
@@ -85,7 +85,10 @@ export function getToolCallComponent({ toolName, toolCallId }: ToolInvocation) {
 
   // Default for other tools
   return (
-    <div key={toolCallId} className="flex items-center gap-1 py-1 px-2 bg-primary/5 rounded-sm border w-fit text-xs">
+    <div
+      key={toolCallId}
+      className="flex items-center gap-1 py-1 px-2 bg-primary/5 rounded-sm border w-fit text-xs"
+    >
       <Loader className="h-3 w-3 animate-spin text-primary" />
       <span>Using {getDisplayToolName(toolName)}</span>
     </div>
@@ -161,7 +164,15 @@ export function getToolResultComponent({
   }
 
   // Default generic result for other tools
-  return <GenericSuccess name={getDisplayToolName(toolName)} message={(result as { message?: string }).message ?? getToolInfo(toolName).message} />;
+  return (
+    <GenericSuccess
+      name={getDisplayToolName(toolName)}
+      message={
+        (result as { message?: string }).message ??
+        getToolInfo(toolName).message
+      }
+    />
+  );
 }
 
 /**
