+++ components/VercelChat/tools/mermaid/MermaidDiagram.tsx
@@ -1,11 +1,11 @@
-'use client';
+"use client";
 
-import React from 'react';
-import MermaidErrorFallback from './MermaidErrorFallback'; // Import the fallback component
-import { GenerateMermaidDiagramResult } from '@/lib/tools/generateMermaidDiagram';
-import { useMermaid } from '@/hooks/useMermaid';
-import handleDownload from '@/lib/utils/download-mermaid-chart';
-import MessageMediaDownloadButton from '../MessageMediaDownloadButton';
+import React from "react";
+import MermaidErrorFallback from "./MermaidErrorFallback"; // Import the fallback component
+import { GenerateMermaidDiagramResult } from "@/lib/tools/generateMermaidDiagram";
+import { useMermaid } from "@/hooks/useMermaid";
+import handleDownload from "@/lib/utils/download-mermaid-chart";
+import MessageMediaDownloadButton from "@/components/VercelChat/MessageMediaDownloadButton";
 
 // Define a type for the global mermaid object, or use any if types aren't installed
 declare global {
@@ -29,10 +29,13 @@ interface MermaidDiagramProps {
  * Renders a Mermaid diagram from a definition string.
  * Dynamically imports the Mermaid library on mount.
  */
-const MermaidDiagramInternal: React.FC<MermaidDiagramProps> = ({ chart, id }) => {
+const MermaidDiagramInternal: React.FC<MermaidDiagramProps> = ({
+  chart,
+  id,
+}) => {
   const { isLibraryLoaded, hasError, containerRef, uniqueId } = useMermaid({
     id,
-    chart
+    chart,
   });
 
   // Conditional rendering: Show fallback on error, otherwise the diagram container
@@ -48,13 +51,20 @@ const MermaidDiagramInternal: React.FC<MermaidDiagramProps> = ({ chart, id }) =>
         id={uniqueId}
         className="mermaid"
         // Hide until library loaded and rendered successfully
-        style={{ visibility: (isLibraryLoaded && !hasError) ? 'visible' : 'hidden', backgroundColor: 'white', border: 'none', padding: '0' }}
+        style={{
+          visibility: isLibraryLoaded && !hasError ? "visible" : "hidden",
+          backgroundColor: "white",
+          border: "none",
+          padding: "0",
+        }}
       >
         {chart}
       </pre>
       {/* Download Diagram Button */}
       {isLibraryLoaded && !hasError && (
-        <MessageMediaDownloadButton onClick={() => handleDownload({ containerRef })} />
+        <MessageMediaDownloadButton
+          onClick={() => handleDownload({ containerRef })}
+        />
       )}
       {/* Top gradient overlay */}
       <div className="absolute z-10 transition-opacity duration-300 opacity-0 group-hover:opacity-100 group-focus:opacity-100 focus-within:opacity-100 end-0 top-0 w-full">
@@ -64,16 +74,22 @@ const MermaidDiagramInternal: React.FC<MermaidDiagramProps> = ({ chart, id }) =>
   );
 };
 
-const MermaidDiagram = ({ result }: { result: GenerateMermaidDiagramResult }) => {
+const MermaidDiagram = ({
+  result,
+}: {
+  result: GenerateMermaidDiagramResult;
+}) => {
   const { content, isError } = result;
 
   if (isError) {
     return <MermaidErrorFallback />;
   }
   const chart = content[0].text;
-  const mermaidDiagramWithoutCodeBlock = chart.replace(/```mermaid\n([\s\S]*?)```/g, '$1').trim();
+  const mermaidDiagramWithoutCodeBlock = chart
+    .replace(/```mermaid\n([\s\S]*?)```/g, "$1")
+    .trim();
 
-  return <MermaidDiagramInternal chart={mermaidDiagramWithoutCodeBlock} />
-}
+  return <MermaidDiagramInternal chart={mermaidDiagramWithoutCodeBlock} />;
+};
 
-export default MermaidDiagram; 
+export default MermaidDiagram;
