+++ components/VercelChat/tools/mermaid/MermaidDiagramSkeleton.tsx
